+++ src/onegov.town/onegov/town/app.py
@@ -113,6 +113,7 @@ class TownApp(Framework):
             'js/intercooler.js',
             'js/underscore.js',
             'js/react.js',
+            'js/stickyfooter.js',
             confirm,
             'js/common.js',
             filters='jsmin',